Recommendation Systems in general are used to give suitable recommendations to users depending upon their liking, age, qualities and other conditional factors. Recommendation systems use a large dataset to recommend entities similar to the input entities. The objective of our project is to provide a list of books similar to the ones that the users have inputted.
This project successfully recommends books for a
given user. It suggests similar books to input books
as well. These featurs of the project allows the user
to quickly find the book he/she might like the most
depending on the aleady given ratings by him/her.
The functionality used has been implemented in such
a way that it wont recommend already rated book to
the user.
One of the unique features of this project is that it
not only recommends top 10 books but also tells the
user the scores depending upon which the books have
been recommended. So, the user can find pattern
in the scores and can think about probable books
depending upon the pattern he/she has found.